:construction: wip(playerhome): begin refactoring into unit components
