+++ frontend/src/components/PlayerHome.jsx
@@ -8,74 +8,37 @@ import { MdErrorOutline } from "react-icons/md";
 import PlaylistCard from "./PlaylistCard";
 import ArtistCard from "./ArtistCard";
 import helpers from "../utils/utilityFunctions.js";
+import { useIsMedia } from "../hooks/useIsMobile.js";
 
 // Constants
 const CURRENT_DATE = "2025-01-23 00:03:46";
 const CURRENT_USER = "gabrielisaacs";
 const DEFAULT_THUMBNAIL = "/thumbnail.png";
 
-// Data transformation functions
-const transformJamendoTrack = (track) => ({
-  id: track.id,
-  title: track.name || track.title,
-  artist: track.artist_name || track.artist,
-  thumbnail: track.image || track.thumbnail || DEFAULT_THUMBNAIL,
-  url: track.audio || track.url,
-  duration: track.duration,
-  likes: `${Math.floor((track.listened || 0) / 1000)}k Plays`,
-});
-
-const transformJamendoArtist = (artist) => ({
-  id: artist.id,
-  name: artist.name,
-  image:
-    artist.image && artist.image.trim() !== ""
-      ? artist.image
-      : `https://usercontent.jamendo.com?type=artist&id=${artist.id}&width=300`,
-  followerCount: artist.sharecount || 0,
-});
-
-const transformJamendoAlbum = (album) => ({
-  id: album.id,
-  title: album.name || album.title,
-  artist: album.user_name || album.artist,
-  thumbnail: album.image || album.thumbnail || DEFAULT_THUMBNAIL,
-  trackCount: album.tracks_count || 0,
-});
-
-const transformJamendoPlaylist = (playlist) => {
-  return {
-    id: playlist.id,
-    title: helpers.capitalize(playlist.title) || "",
-    artist: helpers.capitalize(playlist.artist) || "",
-    thumbnail: `https://usercontent.jamendo.com?type=playlist&id=${playlist.id}&width=300`,
-    creationDate: playlist.creationDate,
-    shareUrl: playlist.shareUrl,
-    shortUrl: playlist.shortUrl,
-    userId: playlist.userId,
-  };
-};
-
 // Loading components
-const SectionLoadingMessage = () => (
+const SectionLoadingMessage = ({ isSmall, isMedium }) => (
   <div className="animate-pulse flex flex-col">
-    <div className="h-10 w-[20rem] bg-neutral-800 rounded-2xl"></div>
-    <div className="grid grid-cols-5 gap-4 mt-[1rem]">
-      {[...Array(5)].map((_, i) => (
+    <div className="h-10 w-9/10 bg-neutral-800 rounded-2xl"></div>
+    <div className="grid grid-cols-1 sm:grid-cols-2 md:grid-cols-3 lg:grid-cols-5 gap-4 mt-[1rem]">
+      {[...Array(isSmall ? 2 : isMedium ? 4 : 5)].map((_, i) => (
         <div
           key={i}
-          className="w-[11.5rem] h-[14rem] bg-neutral-800 mb-4 rounded-3xl"
+          className="aspect-square bg-neutral-800 mb-4 rounded-3xl"
         ></div>
       ))}
     </div>
   </div>
 );
 
-const LoadingMessage = () => (
-  <div className="flex mx-16 h-screen max-w-[60rem]">
-    <div className="flex flex-col mt-[1.75rem] gap-[5rem]">
-      {[...Array(3)].map((_, index) => (
-        <SectionLoadingMessage key={index} />
+const LoadingMessage = ({ isSmall, isMedium }) => (
+  <div className="flex mx-2 md:mx-12 lg:mx-16 min-h-screen max-w-[60rem] py-10">
+    <div className="flex flex-col mt-[1.75rem] gap-[5rem] w-full">
+      {[...Array(isSmall ? 1 : isMedium ? 2 : 3)].map((_, index) => (
+        <SectionLoadingMessage
+          key={index}
+          isSmall={isSmall}
+          isMedium={isMedium}
+        />
       ))}
     </div>
   </div>
@@ -114,67 +77,19 @@ const ContentSection = ({ title, loading, error, data, onRetry, children }) => {
 const PlayerHome = () => {
   const { handleTrackSelect } = usePlayer();
   const navigate = useNavigate();
+  const isSmall = useIsMedia(426);
+  const isMedium = useIsMedia(768);
 
   useEffect(() => {
     window.scrollTo(0, 0);
   }, []);
 
-  // State for pagination
-  const [visibleArtists, setVisibleArtists] = useState(0);
-  const [visibleAlbums, setVisibleAlbums] = useState(0);
-  const [visibleTrending, setVisibleTrending] = useState(0);
-  const [visiblePlaylists, setVisiblePlaylists] = useState(0);
+  useEffect(() => {
+    console.log("isSmall:", isSmall, "isMedium:", isMedium);
+  }, [isSmall, isMedium]);
 
   // Constants
   const cardsPerSet = 5;
-  const trendingCardsPerPage = 12;
-
-  // Use the custom hook for data fetching with caching and retry
-  const {
-    data: artists,
-    loading: loadingArtists,
-    error: artistsError,
-    retry: retryArtists,
-  } = useDataFetching(
-    () => api.getTopArtists({ limit: 20 }),
-    "top-artists",
-    []
-  );
-
-  const {
-    data: albums,
-    loading: loadingAlbums,
-    error: albumsError,
-    retry: retryAlbums,
-  } = useDataFetching(() => api.getAlbums({ limit: 20 }), "albums", []);
-
-  const {
-    data: trendingSongs,
-    loading: loadingTrending,
-    error: trendingError,
-    retry: retryTrending,
-  } = useDataFetching(
-    () => api.getTrendingTracks({ limit: 30 }),
-    "trending-tracks",
-    []
-  );
-
-  const {
-    data: playlists,
-    loading: loadingPlaylists,
-    error: playlistsError,
-    retry: retryPlaylists,
-  } = useDataFetching(() => api.getPlaylists({ limit: 20 }), "playlists", []);
-
-  // Transform the data
-  const transformedArtists = artists ? artists.map(transformJamendoArtist) : [];
-  const transformedAlbums = albums ? albums.map(transformJamendoAlbum) : [];
-  const transformedTrending = trendingSongs
-    ? trendingSongs.map(transformJamendoTrack)
-    : [];
-  const transformedPlaylists = playlists
-    ? playlists.map(transformJamendoPlaylist)
-    : [];
 
   // Navigation handlers
   const handleNext = (setVisible, visible, totalItems) => {
@@ -254,22 +169,6 @@ const PlayerHome = () => {
     }
   };
 
-  const handleArtistClick = (artist) => {
-    window.scrollTo(0, 0);
-    navigate(`/artist/${artist.id}`);
-  };
-
-  const handleAlbumClick = (album) => {
-    window.scrollTo(0, 0);
-    navigate(`/album/${album.id}`);
-  };
-
-  const handlePlaylistClick = (playlist) => {
-    window.scrollTo(0, 0);
-    console.log("CLICKED ON: ", playlist.id);
-    navigate(`/playlist/${playlist.id}`);
-  };
-
   const truncateTitle = (title, maxLength) => {
     if (!title) return "";
     return title.length > maxLength ? `${title.slice(0, maxLength)}...` : title;
@@ -280,7 +179,7 @@ const PlayerHome = () => {
     loadingArtists && loadingAlbums && loadingTrending && loadingPlaylists;
 
   if (isPageLoading) {
-    return <LoadingMessage />;
+    return <LoadingMessage isSmall={isSmall} isMedium={isMedium} />;
   }
 
   // Check if there's no data at all
@@ -292,7 +191,7 @@ const PlayerHome = () => {
 
   if (hasNoData && !isPageLoading) {
     return (
-      <div className="flex justify-center items-center h-[calc(100svh-5.2rem)] w-full border-2 border-orange-300">
+      <div className="flex justify-center items-center h-[calc(100svh-5.2rem)] w-full ">
         <div className="text-neutral-400">No content available</div>
       </div>
     );
@@ -300,246 +199,7 @@ const PlayerHome = () => {
 
   // Main render
   return (
-    <div className="max-w-[60rem] min-h-screen flex flex-col mt-6 mx-16 gap-10 transition-all duration-300">
-      {/* Popular Artists Section */}
-      {transformedArtists.length > 0 && (
-        <div className="flex flex-col mb-10 w-full">
-          <div className="flex flex-row w-full mb-4 items-center">
-            <p className="text-3xl font-extrabold">Popular Artists</p>
-            <div className="ml-auto flex gap-2 items-center">
-              <button
-                onClick={() => {
-                  window.scrollTo({ top: 0, behavior: "smooth" });
-                  navigate("/artists");
-                }}
-                className="bg-transparent hover:bg-[#212121] py-2 px-4 rounded-full border border-neutral-800 text-sm"
-              >
-                More
-              </button>
-              <button
-                onClick={() =>
-                  handlePrevious(setVisibleArtists, visibleArtists)
-                }
-                className="bg-transparent hover:bg-[#212121] p-2 rounded-full border border-neutral-800"
-              >
-                <FaChevronLeft />
-              </button>
-              <button
-                onClick={() =>
-                  handleNext(
-                    setVisibleArtists,
-                    visibleArtists,
-                    transformedArtists.length
-                  )
-                }
-                className="bg-transparent hover:bg-[#212121] p-2 rounded-full border border-neutral-800"
-              >
-                <FaChevronRight />
-              </button>
-            </div>
-          </div>
-          <div className="flex flex-row bg-transparent h-[16rem] w-full gap-4">
-            {transformedArtists
-              .slice(visibleArtists, visibleArtists + cardsPerSet)
-              .map((artist) => (
-                <ArtistCard
-                  key={artist.id}
-                  artist={artist}
-                  onClick={handleArtistClick}
-                  truncateTitle={truncateTitle}
-                />
-              ))}
-          </div>
-        </div>
-      )}
-
-      {/* Albums section */}
-      {transformedAlbums.length > 0 && (
-        <div className="flex flex-col mb-10">
-          <div className="flex flex-row w-full mb-4 items-center">
-            <p className="text-3xl font-extrabold">Albums for you</p>
-            <div className="ml-auto flex gap-2 items-center">
-              <button
-                onClick={() => handlePrevious(setVisibleAlbums, visibleAlbums)}
-                className="bg-transparent hover:bg-[#212121] p-2 rounded-full border border-neutral-800"
-              >
-                <FaChevronLeft />
-              </button>
-              <button
-                onClick={() =>
-                  handleNext(
-                    setVisibleAlbums,
-                    visibleAlbums,
-                    transformedAlbums.length
-                  )
-                }
-                className="bg-transparent hover:bg-[#212121] p-2 rounded-full border border-neutral-800"
-              >
-                <FaChevronRight />
-              </button>
-            </div>
-          </div>
-          <div className="flex flex-row bg-transparent h-[16rem] w-full gap-4 mt-4">
-            {transformedAlbums
-              .slice(visibleAlbums, visibleAlbums + cardsPerSet)
-              .map((album) => (
-                <button
-                  key={album.id}
-                  onClick={() => handleAlbumClick(album)}
-                  className="flex flex-col bg-white bg-opacity-[2%] rounded-xl w-[11.45rem] h-full p-3 gap-4 hover:border-none transition-all relative group hover:bg-opacity-5"
-                >
-                  <div className="opacity-0 group-hover:opacity-100 flex bg-white w-10 h-10 rounded-full shadow-2xl absolute right-6 top-[7.5rem] hover:scale-110 transition-all duration-300">
-                    <FaPlay className="m-auto shadow-lg fill-black" />
-                  </div>
-                  <img
-                    src={album.thumbnail}
-                    className="rounded-xl h-auto w-full shadow-md object-cover"
-                    alt={album.title}
-                  />
-                  <div className="flex flex-col text-left">
-                    <p className="font-bold text-lg">
-                      {truncateTitle(album.title, 12)}
-                    </p>
-                    <p className="font-bold text-sm text-neutral-400">
-                      {truncateTitle(album.artist, 18)}
-                    </p>
-                  </div>
-                </button>
-              ))}
-          </div>
-        </div>
-      )}
-
-      {/* Trending Songs Section */}
-      {transformedTrending.length > 0 && (
-        <div className="flex flex-col mb-10">
-          <div className="flex flex-row w-full mb-4 items-center">
-            <p className="text-3xl font-extrabold">Trending Tracks</p>
-            <div className="ml-auto flex gap-2 items-center transition-all duration-300">
-              <button
-                onClick={handlePlayAll}
-                className="bg-transparent hover:bg-[#212121] py-2 px-4 rounded-full border border-neutral-800 text-sm"
-              >
-                Play all
-              </button>
-              <button
-                onClick={() =>
-                  handlePrevious(setVisibleTrending, visibleTrending)
-                }
-                className="bg-transparent hover:bg-[#212121] p-2 rounded-full border border-neutral-800"
-              >
-                <FaChevronLeft />
-              </button>
-              <button
-                onClick={() =>
-                  handleNext(
-                    setVisibleTrending,
-                    visibleTrending,
-                    transformedTrending.length
-                  )
-                }
-                className="bg-transparent hover:bg-[#212121] p-2 rounded-full border border-neutral-800"
-              >
-                <FaChevronRight />
-              </button>
-            </div>
-          </div>
-          <div className="grid grid-cols-3 gap-2 transition-all duration-300">
-            {transformedTrending
-              .slice(visibleTrending, visibleTrending + trendingCardsPerPage)
-              .map((song, index) => (
-                <button
-                  key={song.id}
-                  onClick={() => handlePlaySong(song, index + visibleTrending)}
-                  className="flex flex-row bg-transparent hover:bg-white hover:bg-opacity-[2%] p-2 rounded-xl gap-3 group text-left transition-all"
-                >
-                  <div className="flex relative">
-                    <img
-                      src={song.thumbnail}
-                      className="h-[3rem] w-[3rem] rounded-lg object-cover"
-                      alt={song.title}
-                    />
-                    <div className="absolute inset-0 flex items-center justify-center opacity-0 group-hover:opacity-100">
-                      <FaPlay className="fill-white drop-shadow-lg" />
-                    </div>
-                  </div>
-                  <div className="flex flex-col flex-1">
-                    <p className="text-base truncate">
-                      {truncateTitle(song.title, 20)}
-                    </p>
-                    <div className="flex flex-row items-center gap-2">
-                      <p className="text-sm opacity-45 truncate">
-                        {song.artist}
-                      </p>
-                      <span className="h-2 w-2 bg-white opacity-45 rounded-full flex-shrink-0"></span>
-                      <p className="text-sm opacity-45 truncate">
-                        {song.likes}
-                      </p>
-                    </div>
-                  </div>
-                </button>
-              ))}
-          </div>
-        </div>
-      )}
-
-      {/* Playlists section */}
-      {transformedPlaylists.length > 0 && (
-        <div className="flex flex-col mb-[10rem]">
-          <div className="flex flex-row w-full mb-4 items-center">
-            <p className="text-3xl font-extrabold">Featured playlists</p>
-            <div className="ml-auto flex gap-2 items-center">
-              <button
-                onClick={() => {
-                  window.scrollTo({ top: 0, behavior: "smooth" });
-                  navigate("/playlists");
-                }}
-                className="bg-transparent hover:bg-[#212121] py-2 px-4 rounded-full border border-neutral-800 text-sm"
-              >
-                More
-              </button>
-              <button
-                onClick={() =>
-                  handlePrevious(setVisiblePlaylists, visiblePlaylists)
-                }
-                className="bg-transparent hover:bg-[#212121] p-2 rounded-full border border-neutral-800"
-              >
-                <FaChevronLeft />
-              </button>
-              <button
-                onClick={() =>
-                  handleNext(
-                    setVisiblePlaylists,
-                    visiblePlaylists,
-                    transformedPlaylists.length
-                  )
-                }
-                className="bg-transparent hover:bg-[#212121] p-2 rounded-full border border-neutral-800"
-              >
-                <FaChevronRight />
-              </button>
-            </div>
-          </div>
-          <div className="flex flex-row bg-transparent h-[16rem] w-full gap-4 mt-4">
-            {transformedPlaylists
-              .slice(visiblePlaylists, visiblePlaylists + cardsPerSet)
-              .map((playlist) => {
-                {
-                  console.log("PLAYLIST MAPPED:", playlist);
-                }
-                return (
-                  <PlaylistCard
-                    key={playlist.id}
-                    playlist={playlist}
-                    onClick={handlePlaylistClick}
-                    truncateTitle={truncateTitle}
-                  />
-                );
-              })}
-          </div>
-        </div>
-      )}
-    </div>
+    <div className="max-w-[60rem] min-h-screen flex flex-col mt-6 mx-16 gap-10 transition-all duration-300"></div>
   );
 };
